test: Move an integration test of compact_candidates_with_memory_budget to the same file
